Update and rename F_W_islands.py to Floyd_Warshall_islands.py
Added normal docstrings
